client/requirements.txt
Flask
Gunicorn
python-dotenv==0.21.0
requests
pyjwt
authlib>=1.0
